+++ dev/modules/MonteCarloSimulationExemplar/source/code/montecarlo_openmp_cpp/coinFlip/coinFlip_seq.cpp
@@ -47,9 +47,9 @@ int main() {
     
 // Try several trials of different numbers of flips, doubling how many each round.
 // 
-// Use a long because we will try a great deal of flips for some trials.
-    long trialFlips = 256;       // start with a small number of flips
-    long maxFlips = 1073741824;  // end with a very large number of flips
+// Use a unsigned int because we will try a great deal of flips for some trials.
+    unsigned int trialFlips = 256;       // start with a small number of flips
+    unsigned int maxFlips = 1073741824;  // end with a very large number of flips
     
     // below we will double the number of trial flips and come back here
     // and run another trial, until we have reached > maxFlips.
@@ -73,7 +73,7 @@ int main() {
         stopTime = clock();   // stop the clock
         
         /***** Show the results  for this trial  *****/
-        printf("%15ld%15d%15d%15d%15.6f%15.6f\n", trialFlips, numHeads, numTails,
+        printf("%15d%15d%15d%15d%15.6f%15.6f\n", trialFlips, numHeads, numTails,
                (numHeads+numTails), chiSq(numHeads, numTails),
                (double)(stopTime-startTime)/CLOCKS_PER_SEC);
